requirements.txt
asgiref==3.6.0
Django==4.2
django-jet-reboot==1.3.3
django-nested-inline==0.4.6
gunicorn==20.1.0
six==1.16.0
sqlparse==0.4.4
